pyproject.toml
[project]
name = "datasets"
version = "0.1.0"
description = "Image processing dataset"
requires-python = ">=3.10"

dependencies = [
    "torch>=1.9.1",              
    "kornia>=0.8.1",             
    "albumentations>=2.0.8",     
    "opencv-python",            
    "numpy",                  
    "pillow",
    "omegaconf",
    "torchvision",
    "hydra-core",    
    "scikit-image>=0.25.2",
    "matplotlib"
]
